requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.2
matplotlib==3.8.4
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
colorama==0.4.6
tqdm==4.66.4
pytest==8.3.2
pytest-cov==5.0.0
flake8==7.1.1
black==24.4.2
mypy==1.10.0
playwright==1.40.0
pytest-playwright==0.4.3
pytest-asyncio==0.21.1
asyncio==3.4.3